+++ R/basics_model_fn.R
@@ -2,10 +2,12 @@
 
 basics_model_fn <- function(train_data,
                             test_data,
-                            fold_info = list(rel_fold=NULL, abs_fold=NULL, fold_column=NULL),
-                            model_specifics = list(model_formula=NULL, family=NULL, link=NULL,
-                                                   control=NULL, REML=FALSE, positive=2,
-                                                   cutoff=0.5, model_verbose=FALSE)){
+                            fold_info = list(rel_fold = NULL, abs_fold = NULL, fold_column = NULL),
+                            model_specifics = list(model_formula = NULL, family = NULL,
+                                                   link = NULL, control = NULL,
+                                                   REML = FALSE, positive = 2,
+                                                   cutoff = 0.5, model_verbose = FALSE,
+                                                   caller = NULL)){
 
   y_col <- extract_y(model_specifics[["model_formula"]]) # Name of target column
   if (is.null(y_col)) stop("The model formula does not contain a dependent variable.")
@@ -23,18 +25,18 @@ basics_model_fn <- function(train_data,
   model_specifics[["model_type"]] <- model_type
 
   # Fit model
-  model <- basics_run_model_fitting(basics_fit_model, model_specifics = model_specifics,
-                                    train_data = train_data,
-                                    warn_info=list(model_formula=model_specifics[["model_formula"]],
-                                                   fold_info=fold_info,
-                                                   model_verbose=model_specifics[["model_verbose"]]))
+  fitted_model <- run_model_fitting(
+    basics_fit_model,
+    model_specifics = model_specifics,
+    train_data = train_data,
+    warn_info = list(
+      model_formula = model_specifics[["model_formula"]],
+      fold_info = fold_info,
+      model_verbose = model_specifics[["model_verbose"]],
+      caller = model_specifics[["caller"]]
+    ))
 
-  if (is.list(model) && isTRUE(model[["yielded_singular_fit_message"]])){
-    yielded_singular_fit_message <- TRUE
-    model <- model[["model"]]
-  } else {
-    yielded_singular_fit_message <- FALSE
-  }
+  model <- fitted_model[["model"]]
 
   # Predict test set
   # If models is NULL (e.g. didn't converge)
@@ -55,9 +57,14 @@ basics_model_fn <- function(train_data,
                                             "abs_fold" = fold_info[["abs_fold"]],
                                             "fold_column" = fold_info[["fold_column"]])
 
-  return(list(predictions_and_targets=predictions_and_targets,
-              model=model,
-              yielded_singular_fit_message = yielded_singular_fit_message))
+  list(
+    predictions_and_targets = predictions_and_targets,
+    model = model,
+    threw_singular_fit_message = fitted_model[["threw_singular_fit_message"]],
+    threw_unknown_message = fitted_model[["threw_unknown_message"]],
+    threw_convergence_warning = fitted_model[["threw_convergence_warning"]],
+    threw_unknown_warning = fitted_model[["threw_unknown_warning"]]
+  )
 
 }