+++ R/cross_validate_fn_single.R
@@ -1,8 +1,8 @@
 # R CMD check NOTE handling
 if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
 
-cross_validate_fn_single <- function(data, model_fn, evaluation_type="linear_regression",
-                                     model_specifics=list(), model_specifics_update_fn=NULL,
+cross_validate_fn_single <- function(data, model_fn, evaluation_type = "gaussian",
+                                     model_specifics = list(), model_specifics_update_fn = NULL,
                                      fold_cols =".folds", parallel_ = FALSE){
 
   # TODO: the below comment is not correct
@@ -28,7 +28,6 @@ cross_validate_fn_single <- function(data, model_fn, evaluation_type="linear_reg
     n_folds <- nlevels(data[[fold_cols]])
   }
 
-
   # Loop through the folds
   # .. Create a test_data and a training_set
   # .. Train the model on the training_set
@@ -53,9 +52,9 @@ cross_validate_fn_single <- function(data, model_fn, evaluation_type="linear_reg
     }
 
     # Create training set for this iteration
-    train_data = data[data[[current_fold_col_name]] != rel_fold,]
+    train_data <- data[data[[current_fold_col_name]] != rel_fold,]
     # Create test set for this iteration
-    test_data = data[data[[current_fold_col_name]] == rel_fold,]
+    test_data <- data[data[[current_fold_col_name]] == rel_fold,]
 
     # Remove folds column(s) from subsets, so we can use "y ~ ." method
     # when defining the model formula.
@@ -69,10 +68,10 @@ cross_validate_fn_single <- function(data, model_fn, evaluation_type="linear_reg
 
     model_fn(train_data = train_data,
              test_data = test_data,
-             fold_info = list(rel_fold=rel_fold,
-                              abs_fold=abs_fold,
-                              fold_column=current_fold_col_name),
-             model_specifics=model_specifics)
+             fold_info = list(rel_fold = rel_fold,
+                              abs_fold = abs_fold,
+                              fold_column = current_fold_col_name),
+             model_specifics = model_specifics)
 
   })
 
@@ -84,11 +83,19 @@ cross_validate_fn_single <- function(data, model_fn, evaluation_type="linear_reg
 
   # Extract models
   models = fold_lists_list %c% 'model'
-  n_conv_warns <- count_nulls_in_list(models)
 
-  # Extract singular fit message
-  singular_fit_messages = fold_lists_list %c% 'yielded_singular_fit_message'
+  # Extract singular fit message flags
+  singular_fit_messages = fold_lists_list %c% 'threw_singular_fit_message'
   n_singular_fit_messages <- sum(unlist(singular_fit_messages))
+
+  # Extract convergence warning flags
+  convergence_warnings = fold_lists_list %c% 'threw_convergence_warning'
+  n_conv_warns <- sum(unlist(convergence_warnings))
+  stopifnot(count_nulls_in_list(models) == n_conv_warns)
+
+  # Extract unknown warning flags
+  unknown_warnings = fold_lists_list %c% 'threw_unknown_warning'
+  n_unknown_warns <- sum(unlist(unknown_warnings))
 
   model_evaluation <- internal_evaluate(
     data = predictions_and_targets,
@@ -105,7 +112,8 @@ cross_validate_fn_single <- function(data, model_fn, evaluation_type="linear_reg
     mutate(Folds = n_folds,
            `Fold Columns` = length(fold_cols),
            `Convergence Warnings` = n_conv_warns,
-           `Singular Fit Messages` = n_singular_fit_messages)
+           `Singular Fit Messages` = n_singular_fit_messages,
+           `Other Warnings` = n_unknown_warns)
 
   return(model_evaluation)